// Twitter/src/constants/enums.ts
export enum TokenType {
  ACCESS_TOKEN,
  REFRESH_TOKEN,
  FORGOT_PASSWORD_TOKEN,
  VERIFY_EMAIL_TOKEN
}

export enum MediaType {
  IMAGE = 'image',
  VIDEO = 'video'
}
export enum TweetAudience {
  EVERYONE = 'everyone',
  CIRCLE = 'circle',
}
export enum TweetType {
  TWEET = 'tweet',
  RETWEET = 'retweet',
  COMMENT = 'comment',
  QUOTE_TWEET = 'quote_tweet'
}